ComboBox: fix empty popup being shown after model is cleared
Currently, ComboBox's popup is sized vertically in this way:

    implicitHeight: contentItem.implicitHeight

Adding an item to a ComboBox with an empty model and then opening
the popup results in the implicitHeight being used in the line below
(in QQuickPopupPositioner::reposition()):

    QRectF rect(p->allowHorizontalMove ? p->x : popupItem->x(),
                p->allowVerticalMove ? p->y : popupItem->y(),
                !p->hasWidth && iw > 0 ? iw : w,
                !p->hasHeight && ih > 0 ? ih : h);

An explicit height was never set on the popup, and ih (the
implicitHeight of the popupItem) is greater than 0. This is fine.

However, when a ComboBox's popup item grows large enough that it has to
be resized to fit within the window, its explicit height is set. The
problem occurs when the model is then cleared, as the implicit height
of the popup item becomes 0. So, while "!p->hasHeight" is still true,
"ih > 0" is not, and the explicit height of the popup item is used,
which is still the previous "let's fill the entire height of the
window" size.

To fix this, we bind the height of the popup to a different expression:

    height: Math.min(contentItem.implicitHeight,
        control.Window.height - topMargin - bottomMargin)

This ensures that the popup has a zero height when the ListView's
implicitHeight is zero (i.e the model is empty), and a height
that fits within the window in all other cases.

Ideally, we'd have a maximumHeight property that controls this, but
for 5.9, we have to fix it this way.

Task-number: QTBUG-60684
Change-Id: Ied94c79bb7b0e693be34e9c7282d991f6f704770
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>