Revert "QQuickRectangle: Optimize setGradient"
This reverts commit 9709d04ba7787c853a1ddbeed0347eab27c0924f.

We changed this from using a signal for notification to using parenting. But
this cannot be done for two related reasons: the first:

    QtObject {
        property Gradient someGradient: Gradient {
            id: someGradient
        }
    }
    ...
    Rectangle {
        gradient: someGradient
    }

... should work, and no longer did, because the Gradient's parent was
the QtObject, not the Rectangle it was tied to. And secondly:

    Rectangle {
        id: rectOne
        gradient: Gradient {
            id: someGradient
        }
    }
    Rectangle {
        id: rectTwo
        gradient: someGradient
    }

... also didn't work anymore, as the Gradient was only notifying a
single Item about changes to it.

The easiest way to fix this is to just revert back to using a signal
connection.

This also caused problems in qtquickcontrols autotests.

Change-Id: Ic07dd02e9920596d0c047bfc23995b3be8c96c49
Task-number: QTBUG-60268
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>