QQuickContainer: removeItem(Item) and takeItem(int)
Deprecate removeItem(int) in favor of removeItem(Item) and
takeItem(int) with clearer semantics.

If one already has access to an item, one can pass it to removeItem()
to get rid of it for good, instead of having to call destroy() by hand.
If one knows the index instead, we provide a way to take the respective
item out of the container, and let the user decide what to do with it.
This way we provide convenient ways to remove & destroy items in one
go, and to take items out without destroying them.

This nice pattern was once upon time planned for the QQC1 containers
such as SplitView, and should have been used for the QQC2 containers,
but was simply forgotten during the development...

[ChangeLog][Controls][Container] Deprecated removeItem(int) in favor of
removeItem(Item) and takeItem(int) with clearer semantics. The former
destroys the item, whereas the latter transfers ownership to the
caller.

Change-Id: I4af51157b196c273f923d99b949a9660e6e8df64
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>