ScrollBar: fix flashing in Material & Universal styles
Ensure a suitable starting opacity for the inactive state transition.

When calling increase() and decrease() from key-handlers, as the docs
suggest, we flash the scrollbar by turning the active state on, and
then back off immediately. If there is an opacity animation when the
scrollbar becomes active (Material & Universal), the opacity animation
is stopped right away, because the state changes back to inactive. This
state changing trick worked only with the Default style, because it
changes the opacity without animating when it becomes active.

Change-Id: I4117de79c7145a710c0b6c43873ca2336b64e21e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>