Expose PaddedRectangle in QtQuick.Controls.impl 2.0
PaddedRectangle was moved from QtQuick.Controls.Material.impl 2.0 to
QtQuick.Controls.impl 2.3 in 8290b876b. It seems that the qhours app
is using the PaddedRectangle for a custom SpinBox. Importing 2.3 would
make Qt 5.10 a hard requirement, so registering PaddedRectangle to
QtQuick.Controls.impl 2.0 allows us to fix qhours so that it still
works with Qt 5.9 and earlier. It's an internal type, and technically
it already was there since 2.0, so this shouldn't matter much.

Change-Id: Ie959dc150356fd5624e899eac9e52ca5a183ffe5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>