Cleanup control flow handling
Replace Loop and ScopeAndFinally with a single ControlFlow
data structure, that can deal with non local control flow.

The control flow data structure can do required cleanup
of internal scopes (for catch or with), emit finally
statements and dispatch to the right continuation.

Also implemented support for the with statement again.

Change-Id: I8bb5414151d0312cb28f8d10e99f8d2b64484100
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>