Manual test: combinations of DragHandler and TapHandler on Flickable
We already had a manual test using a slider, but that has only
one combination of DragHandler and TapHandler.

This test aims to test all possible combinations of DragHandler and
TapHandler together (as siblings, in different parts of the hierarcy,
with a Flickable beneath...)

We also show the current grabbers as an overlay over this entire
collection of manual tests.

Change-Id: Ic634d36d14f7456170f43b077fa72b03fb65bc18
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>