QQmlXMLHttpRequest: support sending ArrayBuffer data
The XMLHttpRequest.send() method should be able to send arbitrary binary
data, and not just UTF-8 text: with this change we first attempt to use
the parameter to the send() method as an ArrayBuffer, and fall back to a
QString if that fails.

[ChangeLog][QtQml] Allow sending binary data, encoded as ArrayBuffer
objects, via XMLHttpRequest's send() method.

Task-number: QTBUG-61599
Change-Id: I25781969ee39b4d168e5c76315ed9853092b322b
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>