Fix tst_calendar::MonthGrid::test_locale()
When the JS Date constructor is passed a string in the ISO 8601 format
(e.g. "1970-01-01"), it is treatead as UTC, not local. Therefore, while
comparing the expected dates from the model, we must base the comparison
on UTC as well. Otherwise, the actual and expected dates may differ due
to the timezone difference. This issue was detected thanks to a new CI
setup that had accidentally UTC+8 timezone set.

Task-number: QTBUG-61535
Change-Id: I93cfce5a81d48dc859f002dcd6b59ddb0701911c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>