Fix bug in designer support reference counting
The QtQuick designer support may override the meta-object and reparent the
property-cache of the object under editing. The code for replacing the parent
of the property cache however was not handling the refcount of the _parent
correctly.

Change-Id: Ic73294fc208b297e8ec9c0b775b5da01a309dba6
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>
(cherry picked from commit 7dc5cd942eb5c55a26d539eedff48b2d6c8e0fc3)
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>