Add QQuickItem::ItemEnabledHasChanged
The itemChange() method has been very useful for Qt Quick Controls 2
to efficiently react to various item changes, but a notification for
the enabled state was missing, so it always had to be handled as a
special case using signals and slots. This change allows QQC2 to handle
enabled state changes the same way e.g. visibility changes are handled.
It's also nice to be able to update a control's internal state before
the actual notifier signal is emitted.

[ChangeLog][QtQuick][QQuickItem] Added a ItemEnabledHasChanged value to
the ItemChange enum. QQuickItem::itemChange(ItemEnabledHasChanged) gets
called when the item's effective enabled state has changed. The new
enabled state is stored in ItemChangeData::boolValue.

Change-Id: Iae96ec21f2b94f453632282473decd1c66097a75
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>