QQuickContainer: use get() instead of QQmlListProperty::data for d-ptr
MSVC2015 has very strange issues with QQmlListProperty::data in release
builds. The pointer is somehow lost between constructing the object and
when it reaches a list-property function. It's probably somehow related
to the qSwap() calls in QQmlObjectCreator::setupBindings(), but this is
tricky to investigate, because adding qDebug() output anywhere near the
qSwap() calls also makes the problem disappear. It's easier to use the
standard Private::get() pattern instead. :)

Change-Id: Iaa361f07023820bfe38191a100cebf64fadb4d62
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>