Fix handling of qsTr & friends
Translation bindings have a special type. Previously, this was done as
a pass on the IR after constant folding: any binding that ended up being
a call to qsTr (and friends) was then changed to being a translation
binding.

Now that we cannot do that anymore, check the AST if the right-hand side
of a script binding is a call to qsTr, and go from there.

Change-Id: Ibd4f6bc947523652fe2319c1adec6f798b952bbf
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>