Fix test for highdpi systems
In this specific case, the original image is rendered at 212x300. If it
is then scales (preserving aspect ratio) to 200x200, the width "should
be" (212/300)*200 = 141.333.. Now when the backing store is not using
highdpi, it will be rendered at 1x, so the width gets rounded to 141.
However, if the backing store renders it at (say) 2x (so width 282.66..)
it gets rounded to 283, which is then divided by 2, which makes 141.5.
By rounding the width down, the result is always the same as on
non-highdpi.

Change-Id: I8c967edf60ddbe97496cfb3d561357887a177d3f
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>