QQuickMenu: use key press instead of key release events
In general, when shortcuts override key press events, the respective
key release events are still sent. Therefore, in order to respect
shortcuts, key navigation should be based on key press events. Key
release -based navigation is also not going to work with nested menus
in a menubar, because a leaf menu may need to propagate key events up
to the parent menus and finally to the menubar, but only if keys are
not actually handled.

Change-Id: Ia040a387adc782d8828250fbed393ef1794f2148
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>