Fix alignment issue on most 32bit arm platforms
The ARM eabi states that 64bit integers are 8-byte aligned (note: iOS
explicitly makes an exception to this). The size of the ExecutionContext
is 9 "pointers" in size, the CallData is 4 "pointers" in size, so on
64bit architectures CallContext is a nice 13 "pointers" sized structure.
However, on ARM the last element of CallData (ValueArray) has a
QV4::Value as its last element, and this needs to be 8-byte aligned. So,
padding will be inserted, and that will grow it from 15 "pointers" to
16.

Task-number: QTBUG-61874
Change-Id: Iba4bc805ef6b459928ef09d604504e896732a8f2
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>