Fix flaky failure in attachedProperty() autotest
QVERIFY(QTest::qWaitForWindowActive(&view)) sometimes failed.

This was because at that point, "view" already had been active, then
turned inactive again due to that the "extraWindow" had become active.
The fix is to not show the "extraWindow" immediately from the QML file,
but to postpone displaying the "extraWindow" to after we have verified
that the "view" has become active.

Change-Id: Ic008a332a736a3b7ab29ad9b2bfeb1eef0d7c19d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>