Fix failing tests in tst_menu and tst_qquickmenubar
As mentioned in the bug report for the failing tst_qquickmenubar test,
the window is sometimes opened at the bottom right corner of the screen
on the Ubuntu CI machines. However, even after centering the window
on the screen, the old setPos() call still caused the tests to fail.
So, we position the cursor past a different corner of the window, which
works.. though I don't know why.

Task-number: QTBUG-62363
Task-number: QTBUG-62383
Change-Id: I3649230002eaac5a6391daff24427e6f9f58cfa5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>