Cleanup argument handling in contexts
Fix the compiler to already deal with duplicated argument names.
Doing this at runtime was not ideal.

Remove the callData member from the context. Instead use the fact that
the arguments already followed the locals in the context. Don't copy the
thisObject over into the CallContext anymore, it's never used from there
anyway.

Fix the ordering of names in the internalclass, so that arguments don't
require special handling anymore when looking them up by name.

Adjust all places that used callData, and related methods.

Change-Id: I0bc45e1be3f1fcd38dc3b4f04e91edaf7b9ed103
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>