QQuickMultiPointHandler::eligiblePoints: use parent not target
If a PinchHandler for example is declared with target: null, it was
crashing here.  The eligible points are contained within the Item
within which the Handler is declared; it's OK to do a pinch gesture
within one Item but manipulate another Item instead.  In such a
case, the parent and target are different.  Alternatively, target
can be set to null, if the user is only interested in reacting to
the Handler's properties in some other way rather than directly
manipulating some target.

Change-Id: Ia077006be1285c242fe7ba71ea89850cd7717c7c
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>