Fix finding test data on Android
The standard way of doing Qt Quick testing requires having
qml files available in the deployment somewhere in the
regular file system. This is not how Android works, where
all test data should be deployed as resources. To allow
this, we simply hardcode the test path as a resource path
on Android instead of checking on the disk.

Task-number: QTBUG-45576
Change-Id: I0729f5004a6206d2a10b782eb5b7d83a833c4f51
Reviewed-by: BogDan Vatra <bogdan@kdab.com>