SwipeDelegate: grab the mouse when accepting a press event
After 1f4bfc099b7e48e2cc3dd9488b7b42a96122b299 in qtdeclarative,
QQuickWindow does not cause the Item to which it's about to attempt
to deliver an event to pre-grab.  (That caused a lot of thrashing
of the grabber, so not doing that is a nice simplification, and
also avoids the need to ungrab later, and also makes mouse event
handling more like touch event handling.)  So whenever SwipeDelegate
knows at the time of mouse press that it wants to see the release
too, it should grab.  The grabber can be either the Item for which
it's filtering events, or the SwipeDelegate itself, whichever is
more appropriate in any given scenario.  For now I assume it should
be the delivery-target Item, because that's what would have had the
grab before.

Task-number: QTBUG-62412
Task-number: QTBUG-62631
Change-Id: Ie01276508c6602b10f9d7996a748ffe90efa36b5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>