sendFilteredPointerEvent: go back to parent traversal
The filteringParentItems vector-of-pairs is getting more and more
questionable: it needs to have more pairs than we thought, then we
need to remove them to prevent multiple-filtering, and then we also
need to keep track of which ones we removed plus which ones are
actually filtering, in order to avoid direct delivery of pointer events
to the same Item after it already filtered from its children.
The overhead doesn't seem worthwhile anymore.  So this conceptually
reverts 9b5fc80af28580e9672792dd511d876a93947882 but with some
improvements.

Also, do not deliver events to child-filtering items which have
already filtered.

Change-Id: I9d4f977dba695de7eb78ab536e0e6e8fd6a253a7
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>