failsafe: never allow touch or mouse grab to persist after release
If a PointerHandler acquires a passive grab and we fail to remove it,
then next time the mouse moves, a filtering parent Flickable will have
a chance to see the mouse hover movement in the course of delivery
to that handler, and will grab.  Thus you are stuck in flicking state
even though no mouse button is held.  Passive grabs must therefore
always be cleared on release.

Change-Id: Ibaf46de32d7855e7b5f377c2d4fdae1312601c4f
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>