ColorImage: colorize only if needed
QQuickColorImageProvider got replaced by QQuickColorImage in 9fc3659.
This seems to have introduced a performance regression in several
controls that are now colorizing images even when there's no need to.

Various indicator images have the correct default color that matches
the normal state. They need to be colorized only when gaining focus
or being disabled, for example.

The improvement in qmlbench results:
- CheckBox: 70 => 84 frames
- ComboBox: 88 => 92 frames
- Dial: 73 => 77 frames
- MenuItem: 73 => 84 frames

Change-Id: I9155042542f5069cff201a1730b9dd2a62cffd67
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>