Fix QNX build on Windows
Configure on Windows in Qt 5.6 does not detect the target gcc version
correctly and reports the host gcc version instead. That means we'll end
up enabling -fno-lifetime-dse with a gcc version that doesn't support it
and consequently fail to build.

Since in the 5.6 branch we only support QNX 6 and we know that the gcc
version is 4.7 and thus not affected by the DSE problem, we might as
well unconditionally disable the workaround there.

This change is only for 5.6 as 5.8/5.9 and newer detect the gcc version
correctly _and_ don't need the -fno-lifetime-dse workaround anymore.

Task-number: QTBUG-62820
Change-Id: I55baf532a9126eb2f8c5f11858d52cccad6c355b
Reviewed-by: Sami Nurmenniemi <sami.nurmenniemi@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>